+++ domain/photo/photo.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Photo struct {
-	ID        int64      `json:"id" db:"id"`
+	ID        int        `json:"id" db:"id"`
 	Title     string     `json:"title" db:"title"`
 	Caption   string     `json:"caption" db:"caption"`
 	PhotoUrl  string     `json:"photoUrl" db:"photo_url"`
-	UserID    int64      `json:"userId" db:"user_id"`
+	UserID    int        `json:"userId" db:"user_id"`
 	CreatedAt time.Time  `json:"-" db:"created_at"`
 	UpdatedAt *time.Time `json:"-" db:"updated_at"`
 }
@@ -24,32 +24,32 @@ type CreatePhoto struct {
 	Title    string `json:"title"`
 	Caption  string `json:"caption"`
 	PhotoUrl string `json:"photo_url"`
-	UserID   int64  `json:"user_id"`
+	UserID   int    `json:"user_id"`
 }
 
 type UpdatePhoto struct {
-	ID       int64  `json:"id"`
+	ID       int    `json:"id"`
 	Title    string `json:"title"`
 	Caption  string `json:"caption"`
 	PhotoUrl string `json:"photo_url"`
-	UserID   int64  `json:"user_id"`
+	UserID   int    `json:"user_id"`
 }
 
 type CreatePhotoResponse struct {
-	ID        int64     `json:"id"`
+	ID        int       `json:"id"`
 	Title     string    `json:"title"`
 	Caption   string    `json:"caption"`
 	PhotoUrl  string    `json:"photo_url"`
-	UserID    int64     `json:"user_id"`
+	UserID    int       `json:"user_id"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
 type PhotoResponse struct {
-	ID        int64      `json:"id"`
+	ID        int        `json:"id"`
 	Title     string     `json:"title"`
 	Caption   string     `json:"caption"`
 	PhotoUrl  string     `json:"photo_url"`
-	UserID    int64      `json:"user_id"`
+	UserID    int        `json:"user_id"`
 	CreatedAt time.Time  `json:"created_at"`
 	UpdatedAt *time.Time `json:"updated_at"`
 	User      UserPhoto  `json:"User"`
@@ -61,10 +61,10 @@ type UserPhoto struct {
 }
 
 type UpdatePhotoResponse struct {
-	ID        int64      `json:"id"`
+	ID        int        `json:"id"`
 	Title     string     `json:"title"`
 	Caption   string     `json:"caption"`
 	PhotoUrl  string     `json:"photo_url"`
-	UserID    int64      `json:"user_id"`
+	UserID    int        `json:"user_id"`
 	UpdatedAt *time.Time `json:"updated_at"`
 }
